Remove enzyme tests
