+++ frontend/lib/src/components/widgets/FileUploader/FileUploader.test.tsx
@@ -15,11 +15,10 @@
  */
 
 import "@testing-library/jest-dom"
-import { screen } from "@testing-library/react"
-import { ReactWrapper, ShallowWrapper } from "enzyme"
+import { fireEvent, screen, waitFor, within } from "@testing-library/react"
 import React from "react"
-import { FileError } from "react-dropzone"
-import { mount, shallow, render } from "@streamlit/lib/src/test_util"
+import { render } from "@streamlit/lib/src/test_util"
+import userEvent from "@testing-library/user-event"
 
 import {
   FileUploader as FileUploaderProto,
@@ -30,17 +29,10 @@ import {
   IFileURLs,
 } from "@streamlit/lib/src/proto"
 import { WidgetStateManager } from "@streamlit/lib/src/WidgetStateManager"
-import FileDropzone from "./FileDropzone"
 import FileUploader, { Props } from "./FileUploader"
-import {
-  ErrorStatus,
-  UploadFileInfo,
-  UploadedStatus,
-  UploadingStatus,
-} from "./UploadFileInfo"
 
-const createFile = (): File => {
-  return new File(["Text in a file!"], "filename.txt", {
+const createFile = (filename = "filename.txt"): File => {
+  return new File(["Text in a file!"], filename, {
     type: "text/plain",
     lastModified: 0,
   })
@@ -55,26 +47,11 @@ const buildFileUploaderStateProto = (
         new UploadedFileInfoProto({
           fileId: fileUrls.fileId,
           fileUrls,
-          name: "filename.txt",
+          name: fileUrls.fileId,
           size: 15,
         })
     ),
   })
-
-const INVALID_TYPE_ERROR: FileError = {
-  message: "error message",
-  code: "file-invalid-type",
-}
-
-const TOO_MANY_FILES: FileError = {
-  message: "error message",
-  code: "too-many-files",
-}
-
-const FILE_TOO_LARGE: FileError = {
-  message: "error message",
-  code: "file-too-large",
-}
 
 const getProps = (elementProps: Partial<FileUploaderProto> = {}): Props => {
   return {
@@ -111,18 +88,7 @@ const getProps = (elementProps: Partial<FileUploaderProto> = {}): Props => {
   }
 }
 
-/** Return a strongly-typed wrapper.state("files") */
-function getFiles(
-  wrapper: ShallowWrapper<FileUploader> | ReactWrapper<FileUploader>,
-  statusType?: string
-): UploadFileInfo[] {
-  const result = wrapper.state<UploadFileInfo[]>("files")
-  return statusType != null
-    ? result.filter(f => f.status.type === statusType)
-    : result
-}
-
-describe("FileUploader widget RTL tests", () => {
+describe("FileUploader widget tests", () => {
   it("renders without crashing", () => {
     const props = getProps()
     render(<FileUploader {...props} />)
@@ -186,106 +152,194 @@ describe("FileUploader widget RTL tests", () => {
     expect(labelNode).toBeInTheDocument()
     expect(labelNode).not.toBeVisible()
   })
-})
 
-// TODO(vdonato / kajarenc): Eventually convert all of these tests to use RTL.
-// Ideally, we'd like to migrate off enzyme while we're working on the big
-// `st.file_uploader` refactor, but migration of a lot of our existing tests
-// would be very involved. Thus, we'll punt on the especially difficult tests
-// for now and revisit rewriting them after the feature branch has landed on
-// `develop`.
-describe("FileUploader widget enzyme tests", () => {
-  it("uploads a single selected file", async () => {
+  it("uploads a single file upload", async () => {
+    const user = userEvent.setup()
     const props = getProps()
     jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
-    const wrapper = shallow(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
-    const fileDropzone = wrapper.find(FileDropzone)
-    fileDropzone.props().onDrop([createFile()], [])
+    render(<FileUploader {...props} />)
 
-    // Wait one tick to simulate FileURLsProto being received from the server,
-    // after which we begin the file upload.
-    await process.nextTick
+    const fileDropZoneInput = screen.getByTestId(
+      "stDropzoneInput"
+    ) as HTMLInputElement
 
-    // We should have 1 file in the uploading state
-    expect(props.uploadClient.uploadFile).toHaveBeenCalledTimes(1)
-    expect(getFiles(wrapper).length).toBe(1)
-    expect(getFiles(wrapper)[0].status.type).toBe("uploading")
-    expect(instance.status).toBe("updating")
+    const fileToUpload = createFile()
 
-    // WidgetStateManager should have been called on mounting
-    expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledTimes(1)
+    await user.upload(fileDropZoneInput, fileToUpload)
 
-    // Wait a tick to simulate the file upload completing.
-    await process.nextTick
+    const fileName = screen.getByTestId("stUploadedFile")
+    expect(fileName.textContent).toContain("filename.txt")
+    expect(fileDropZoneInput.files?.[0]).toEqual(fileToUpload)
 
-    // After upload completes, our file should be "uploaded" and our status
-    // should be "ready". The file should also have a new ID, returned by our
-    // uploadFile endpoint.
-    const files = getFiles(wrapper)
-    expect(files.length).toBe(1)
-    const fileStatus = files[0].status as UploadedStatus
-    expect(fileStatus.type).toBe("uploaded")
-    expect(fileStatus.fileId).toBe("filename.txt")
-    expect(fileStatus.fileUrls).toEqual({
-      deleteUrl: "filename.txt",
-      fileId: "filename.txt",
-      uploadUrl: "filename.txt",
-    })
-    expect(instance.status).toBe("ready")
-
-    // And WidgetStateManager should have been called with the file's ID
     expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledWith(
       props.element,
-      buildFileUploaderStateProto([fileStatus.fileUrls]),
+      buildFileUploaderStateProto([
+        {
+          fileId: "filename.txt",
+          uploadUrl: "filename.txt",
+          deleteUrl: "filename.txt",
+        },
+      ]),
       {
         fromUi: true,
       }
     )
   })
-
   it("uploads a single file even if too many files are selected", async () => {
     const props = getProps({ multipleFiles: false })
     jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
-    const wrapper = shallow(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
-    const fileDropzone = wrapper.find(FileDropzone)
-    fileDropzone.props().onDrop(
-      [],
-      [
-        {
-          file: createFile(),
-          errors: [INVALID_TYPE_ERROR, TOO_MANY_FILES],
-        },
-        { file: createFile(), errors: [TOO_MANY_FILES] },
-        { file: createFile(), errors: [TOO_MANY_FILES] },
-      ]
+    render(<FileUploader {...props} />)
+
+    const fileDropZone = screen.getByTestId(
+      "stFileUploadDropzone"
+    ) as HTMLElement
+
+    const filesToUpload = [
+      new File(["Text in a file!"], "filename1.txt", {
+        type: "text/plain",
+        lastModified: 0,
+      }),
+      new File(["Text in an another file!"], "filename2.txt", {
+        type: "text/plain",
+        lastModified: 0,
+      }),
+      new File(["Another text in an another file!"], "filename3.txt", {
+        type: "text/plain",
+        lastModified: 0,
+      }),
+    ]
+
+    fireEvent.drop(fileDropZone, {
+      dataTransfer: {
+        types: ["Files", "Files", "Files"],
+        files: filesToUpload,
+      },
+    })
+
+    await waitFor(() =>
+      expect(props.uploadClient.uploadFile).toHaveBeenCalledTimes(1)
     )
 
-    await process.nextTick
-    expect(props.uploadClient.uploadFile).toHaveBeenCalledTimes(1)
-
+    const fileElements = screen.getAllByTestId("stUploadedFile")
     // We should have 3 files. One will be uploading, the other two will
     // be in the error state.
-    expect(getFiles(wrapper).length).toBe(3)
-    expect(getFiles(wrapper, "uploading").length).toBe(1)
-    expect(getFiles(wrapper, "error").length).toBe(2)
-    expect(instance.status).toBe("updating")
+    expect(fileElements.length).toBe(3)
+    expect(fileElements[0].textContent).toContain("filename1.txt")
 
-    await process.nextTick
+    const errors = screen.getAllByTestId("stUploadedFileErrorMessage")
 
-    // Our upload has completed, our file should now be "uploaded", and
-    // our status should now be "ready".
-    expect(getFiles(wrapper).length).toBe(3)
-    expect(getFiles(wrapper, "uploaded").length).toBe(1)
-    expect(getFiles(wrapper, "error").length).toBe(2)
-    expect(instance.status).toBe("ready")
+    expect(errors.length).toBe(2)
+    expect(errors[0].textContent).toContain("Only one file is allowed.")
+    expect(errors[1].textContent).toContain("Only one file is allowed.")
 
-    const uploadedFile = getFiles(wrapper, "uploaded")[0]
     expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledWith(
       props.element,
       buildFileUploaderStateProto([
-        (uploadedFile.status as UploadedStatus).fileUrls,
+        {
+          fileId: "filename1.txt",
+          uploadUrl: "filename1.txt",
+          deleteUrl: "filename1.txt",
+        },
+      ]),
+      {
+        fromUi: true,
+      }
+    )
+  })
+  it("replaces file on single file uploader", async () => {
+    const user = userEvent.setup()
+    const props = getProps()
+    jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
+    render(<FileUploader {...props} />)
+
+    const fileDropZoneInput = screen.getByTestId(
+      "stDropzoneInput"
+    ) as HTMLInputElement
+
+    const firstFile = createFile()
+
+    await user.upload(fileDropZoneInput, firstFile)
+
+    const fileName = screen.getByTestId("stUploadedFile")
+    expect(fileName.textContent).toContain("filename.txt")
+    expect(fileDropZoneInput.files?.[0]).toEqual(firstFile)
+
+    expect(props.uploadClient.uploadFile).toHaveBeenCalledTimes(1)
+
+    const secondFile = new File(["Another text in a file"], "filename2.txt", {
+      type: "text/plain",
+      lastModified: 0,
+    })
+
+    // Upload a replacement file
+    await user.upload(fileDropZoneInput, secondFile)
+
+    const currentFiles = screen.getAllByTestId("stUploadedFile")
+    expect(currentFiles.length).toBe(1)
+    expect(currentFiles[0].textContent).toContain("filename2.txt")
+    expect(fileDropZoneInput.files?.[0]).toEqual(secondFile)
+    expect(props.uploadClient.uploadFile).toHaveBeenCalledTimes(2)
+  })
+
+  it("uploads multiple files, even if some have errors", async () => {
+    const props = getProps({ multipleFiles: true, type: [".txt"] })
+    jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
+    render(<FileUploader {...props} />)
+
+    const fileDropZone = screen.getByTestId(
+      "stFileUploadDropzone"
+    ) as HTMLElement
+
+    const filesToUpload = [
+      new File(["Text in a file!"], "filename1.txt", {
+        type: "text/plain",
+        lastModified: 0,
+      }),
+      new File(["Text in a file?"], "filename2.txt", {
+        type: "text/plain",
+        lastModified: 0,
+      }),
+      new File(["Another PDF file"], "anotherpdffile.pdf", {
+        type: "application/pdf",
+        lastModified: 0,
+      }),
+    ]
+
+    fireEvent.drop(fileDropZone, {
+      dataTransfer: {
+        types: ["Files"],
+        files: filesToUpload,
+        items: filesToUpload.map(file => ({
+          kind: "file",
+          type: file.type,
+          getAsFile: () => file,
+        })),
+      },
+    })
+
+    await waitFor(() =>
+      expect(props.uploadClient.uploadFile).toHaveBeenCalledTimes(2)
+    )
+
+    const fileNames = screen.getAllByTestId("stUploadedFile")
+    expect(fileNames.length).toBe(3)
+
+    const errorFileNames = screen.getAllByTestId("stUploadedFileErrorMessage")
+    expect(errorFileNames.length).toBe(1)
+
+    expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledWith(
+      props.element,
+      buildFileUploaderStateProto([
+        {
+          fileId: "filename1.txt",
+          uploadUrl: "filename1.txt",
+          deleteUrl: "filename1.txt",
+        },
+        {
+          fileId: "filename2.txt",
+          uploadUrl: "filename2.txt",
+          deleteUrl: "filename2.txt",
+        },
       ]),
       {
         fromUi: true,
@@ -293,148 +347,69 @@ describe("FileUploader widget enzyme tests", () => {
     )
   })
 
-  it("replaces file on single file uploader", async () => {
-    const props = getProps({ multipleFiles: false })
-    const wrapper = mount(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
-    const fileDropzone = wrapper.find(FileDropzone)
-
-    // Upload the first file
-    fileDropzone.props().onDrop([createFile()], [])
-    await process.nextTick
-
-    const origFile = getFiles(wrapper)[0]
-    expect(props.uploadClient.uploadFile).toHaveBeenCalledTimes(1)
-    expect(getFiles(wrapper).length).toBe(1)
-    expect(getFiles(wrapper, "uploading").length).toBe(1)
-
-    await process.nextTick
-
-    expect(getFiles(wrapper, "uploaded").length).toBe(1)
-    expect(instance.status).toBe("ready")
-
-    // Upload a replacement file
-    fileDropzone.props().onDrop([createFile()], [])
-    await process.nextTick
-
-    const replacementFile = getFiles(wrapper)[0]
-    expect(props.uploadClient.uploadFile).toHaveBeenCalledTimes(2)
-    expect(getFiles(wrapper).length).toBe(1)
-    expect(getFiles(wrapper, "uploading").length).toBe(1)
-
-    // Ensure that our new UploadFileInfo has a different ID than the first
-    // file's.
-    expect(replacementFile.id).not.toBe(origFile.id)
-
-    await process.nextTick
-
-    // Ensure that our final status is as expected.
-    expect(getFiles(wrapper, "uploaded").length).toBe(1)
-    expect(instance.status).toBe("ready")
-  })
-
-  it("uploads multiple files, even if some have errors", async () => {
-    const props = getProps({ multipleFiles: true })
-    jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
-    const wrapper = shallow(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
-    const fileDropzone = wrapper.find(FileDropzone)
-
-    // Drop two valid files, and 3 invalid ones.
-    fileDropzone.props().onDrop(
-      [createFile(), createFile()],
-      [
-        { file: createFile(), errors: [INVALID_TYPE_ERROR] },
-        { file: createFile(), errors: [INVALID_TYPE_ERROR] },
-        { file: createFile(), errors: [INVALID_TYPE_ERROR] },
-      ]
-    )
-
-    await process.nextTick
-
-    expect(props.uploadClient.uploadFile).toHaveBeenCalledTimes(2)
-
-    // We should have two files uploading, and 3 showing an error.
-    expect(getFiles(wrapper).length).toBe(5)
-    expect(getFiles(wrapper, "uploading").length).toBe(2)
-    expect(getFiles(wrapper, "error").length).toBe(3)
-    expect(instance.status).toBe("updating")
-
-    await process.nextTick
-
-    // When our upload completes, our uploading files should now be uploaded
-    expect(getFiles(wrapper).length).toBe(5)
-    expect(getFiles(wrapper, "uploaded").length).toBe(2)
-    expect(getFiles(wrapper, "error").length).toBe(3)
-    expect(instance.status).toBe("ready")
-
-    // WidgetStateManager should have been called with the server IDs
-    // of the uploaded files.
-    const uploadedFiles = getFiles(wrapper, "uploaded")
-    expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledWith(
-      props.element,
-      buildFileUploaderStateProto(
-        uploadedFiles.map(f => (f.status as UploadedStatus).fileUrls)
-      ),
-      {
-        fromUi: true,
-      }
-    )
-  })
-
   it("can delete completed upload", async () => {
+    const user = userEvent.setup()
     const props = getProps({ multipleFiles: true })
     jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
-    const wrapper = mount(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
+    render(<FileUploader {...props} />)
+
+    const fileDropZoneInput = screen.getByTestId(
+      "stDropzoneInput"
+    ) as HTMLInputElement
 
     // Upload two files
-    // @ts-expect-error -- dummy value being used as first arg
-    instance.uploadFile("someUploadUrl1", createFile())
-    // @ts-expect-error -- dummy value being used as first arg
-    instance.uploadFile("someUploadUrl2", createFile())
+    await user.upload(fileDropZoneInput, createFile("filename1.txt"))
+    await user.upload(fileDropZoneInput, createFile("filename2.txt"))
 
-    await process.nextTick
+    const fileNames = screen.getAllByTestId("stUploadedFile")
+    expect(fileNames.length).toBe(2)
+    expect(fileNames[0].textContent).toContain("filename2.txt")
+    expect(fileNames[1].textContent).toContain("filename1.txt")
 
-    const initialFiles = getFiles(wrapper)
-    expect(initialFiles.length).toBe(2)
-    expect(getFiles(wrapper, "uploaded").length).toBe(2)
-    expect(instance.status).toBe("ready")
-
-    // WidgetStateManager should have been called with our two file IDs
-    expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledTimes(2)
+    // WidgetStateManager should have been called with our two file IDs and first time with empty state
+    expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledTimes(3)
 
     expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenLastCalledWith(
       props.element,
-      buildFileUploaderStateProto(
-        initialFiles.map(f => (f.status as UploadedStatus).fileUrls)
-      ),
+      buildFileUploaderStateProto([
+        {
+          fileId: "filename1.txt",
+          uploadUrl: "filename1.txt",
+          deleteUrl: "filename1.txt",
+        },
+        {
+          fileId: "filename2.txt",
+          uploadUrl: "filename2.txt",
+          deleteUrl: "filename2.txt",
+        },
+      ]),
       {
         fromUi: true,
       }
     )
 
-    // Delete the first file
-    instance.deleteFile(initialFiles[0].id)
+    const firstDeleteBtn = screen.getAllByTestId("fileDeleteBtn")[0]
 
-    await process.nextTick
+    await user.click(within(firstDeleteBtn).getByRole("button"))
 
-    // We should only have a single file - the second file from the original
-    // upload.
-    expect(getFiles(wrapper).length).toBe(1)
-    expect(getFiles(wrapper)[0]).toBe(initialFiles[1])
-    expect(instance.status).toBe("ready")
+    // We should only have a single file - the second file from the original upload list (filename1.txt).
+    const fileNamesAfterDelete = screen.getAllByTestId("stUploadedFile")
+    expect(fileNamesAfterDelete.length).toBe(1)
+    expect(fileNamesAfterDelete[0].textContent).toContain("filename1.txt")
 
     // WidgetStateManager should have been called with the file ID
-    // of the remaining file. This should be the third time WidgetStateManager
+    // of the remaining file. This should be the fourth time WidgetStateManager
     // has been updated.
-    expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledTimes(3)
-    const newWidgetValue = [initialFiles[1]]
+    expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledTimes(4)
     expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenLastCalledWith(
       props.element,
-      buildFileUploaderStateProto(
-        newWidgetValue.map(f => (f.status as UploadedStatus).fileUrls)
-      ),
+      buildFileUploaderStateProto([
+        {
+          fileId: "filename1.txt",
+          uploadUrl: "filename1.txt",
+          deleteUrl: "filename1.txt",
+        },
+      ]),
       {
         fromUi: true,
       }
@@ -442,27 +417,33 @@ describe("FileUploader widget enzyme tests", () => {
   })
 
   it("can delete in-progress upload", async () => {
-    const props = getProps({ multipleFiles: true })
+    const user = userEvent.setup()
+    const props = getProps()
+
+    // Mock the uploadFile method to return a promise that never resolves to test updating state
+    props.uploadClient.uploadFile = jest.fn().mockImplementation(() => {
+      return new Promise(() => {})
+    })
+
     jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
-    const wrapper = mount(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
+    render(<FileUploader {...props} />)
 
-    // Upload a file...
-    // @ts-expect-error -- dummy value being used as first arg
-    instance.uploadFile("someFileUrl", createFile())
-    expect(getFiles(wrapper, "uploading").length).toBe(1)
-    expect(instance.status).toBe("updating")
+    const fileDropZoneInput = screen.getByTestId(
+      "stDropzoneInput"
+    ) as HTMLInputElement
 
-    const fileId = getFiles(wrapper)[0].id
+    await user.upload(fileDropZoneInput, createFile())
+
+    const progressBar = screen.getByRole("progressbar")
+    expect(progressBar).toBeInTheDocument()
 
     // and then immediately delete it before upload "completes"
-    instance.deleteFile(fileId)
+    const deleteBtn = screen.getByTestId("fileDeleteBtn")
 
-    // Wait for the update
-    await process.nextTick
+    await user.click(within(deleteBtn).getByRole("button"))
 
-    expect(getFiles(wrapper).length).toBe(0)
-    expect(instance.status).toBe("ready")
+    const fileNames = screen.queryAllByTestId("stUploadedFile")
+    expect(fileNames.length).toBe(0)
 
     // WidgetStateManager will still have been called once, during component mounting
     expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledTimes(1)
@@ -475,119 +456,132 @@ describe("FileUploader widget enzyme tests", () => {
     )
   })
 
-  it("can delete file with ErrorStatus", () => {
-    const props = getProps()
-    jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
-    const wrapper = shallow(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
-    const fileDropzone = wrapper.find(FileDropzone)
+  it("can delete file with ErrorStatus", async () => {
+    const user = userEvent.setup()
+    const props = getProps({ multipleFiles: false, type: [".txt"] })
+    render(<FileUploader {...props} />)
 
-    // Drop a file with an error status
-    fileDropzone
-      .props()
-      .onDrop([], [{ file: createFile(), errors: [INVALID_TYPE_ERROR] }])
+    const fileDropZone = screen.getByTestId(
+      "stFileUploadDropzone"
+    ) as HTMLElement
 
-    expect(getFiles(wrapper, "error").length).toBe(1)
+    const filesToUpload = [
+      new File(["Another PDF file"], "anotherpdffile.pdf", {
+        type: "application/pdf",
+        lastModified: 0,
+      }),
+    ]
+
+    // Drop a file with an error (wrong extension)
+    fireEvent.drop(fileDropZone, {
+      dataTransfer: {
+        types: ["Files"],
+        files: filesToUpload,
+        items: filesToUpload.map(file => ({
+          kind: "file",
+          type: file.type,
+          getAsFile: () => file,
+        })),
+      },
+    })
+
+    await waitFor(() =>
+      expect(screen.getAllByTestId("stUploadedFile").length).toBe(1)
+    )
+
+    const errorFileNames = screen.getAllByTestId("stUploadedFileErrorMessage")
+    expect(errorFileNames.length).toBe(1)
 
     // Delete the file
-    instance.deleteFile(getFiles(wrapper)[0].id)
+    const firstDeleteBtn = screen.getAllByTestId("fileDeleteBtn")[0]
+
+    await user.click(within(firstDeleteBtn).getByRole("button"))
 
     // File should be gone
-    expect(getFiles(wrapper).length).toBe(0)
+    const fileNamesAfterDelete = screen.queryAllByTestId("stUploadedFile")
+    expect(fileNamesAfterDelete.length).toBe(0)
   })
 
   it("handles upload error", async () => {
+    const user = userEvent.setup()
     const props = getProps()
-    const wrapper = shallow(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
+    jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
+    render(<FileUploader {...props} />)
+
+    const fileDropZoneInput = screen.getByTestId(
+      "stDropzoneInput"
+    ) as HTMLInputElement
 
     // Upload a file that will be rejected by the server
     props.uploadClient.uploadFile = jest
       .fn()
       .mockRejectedValue(new Error("random upload error!"))
 
-    // @ts-expect-error -- dummy value being used as first arg
-    instance.uploadFile("someFileUploadUrl", createFile())
+    await user.upload(fileDropZoneInput, createFile())
 
-    expect(getFiles(wrapper)[0].status.type).toBe("uploading")
-
-    // Wait one tick for the upload to be rejected
-    await process.nextTick
-
-    // And another for the uploader to be re-rendered
-    await process.nextTick
-
-    // Our file should have an error status, and our uploader should still be
-    // "ready"
-    expect(getFiles(wrapper)[0].status.type).toBe("error")
-    expect(
-      (getFiles(wrapper)[0].status as ErrorStatus).errorMessage
-    ).toContain("random upload error!")
-
-    expect(instance.status).toBe("ready")
+    // Our file should have an error status
+    const errorFileNames = screen.getByTestId("stUploadedFileErrorMessage")
+    expect(errorFileNames.textContent).toContain("random upload error!")
   })
 
-  it("changes status + adds file attributes when dropping a File", async () => {
-    const props = getProps()
-    const wrapper = shallow(<FileUploader {...props} />)
-    const fileDropzone = wrapper.find(FileDropzone)
-    fileDropzone.props().onDrop([createFile()], [])
-    await process.nextTick
+  it("shows an ErrorStatus when File extension is not allowed", async () => {
+    const props = getProps({ multipleFiles: false, type: [".png"] })
+    render(<FileUploader {...props} />)
 
-    const files = getFiles(wrapper)
-    expect(files[0].status.type).toBe("uploading")
-    expect((files[0].status as UploadingStatus).cancelToken).toBeDefined()
-    expect(files[0].id).toBeDefined()
-  })
+    const fileDropZone = screen.getByTestId(
+      "stFileUploadDropzone"
+    ) as HTMLElement
 
-  it("shows an ErrorStatus when File extension is not allowed", () => {
-    const props = getProps({ type: ["png"] })
-    const wrapper = shallow(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
-    const fileDropzone = wrapper.find(FileDropzone)
-    fileDropzone
-      .props()
-      .onDrop([], [{ file: createFile(), errors: [INVALID_TYPE_ERROR] }])
+    const filesToUpload = [
+      new File(["TXT file"], "txtfile.txt", {
+        type: "text/plain",
+        lastModified: 0,
+      }),
+    ]
 
-    expect(getFiles(wrapper)[0].status.type).toBe("error")
-    expect((getFiles(wrapper)[0].status as ErrorStatus).errorMessage).toBe(
+    // Drop a file with an error (wrong extension)
+    fireEvent.drop(fileDropZone, {
+      dataTransfer: {
+        types: ["Files"],
+        files: filesToUpload,
+        items: filesToUpload.map(file => ({
+          kind: "file",
+          type: file.type,
+          getAsFile: () => file,
+        })),
+      },
+    })
+
+    await waitFor(() =>
+      expect(screen.getAllByTestId("stUploadedFile").length).toBe(1)
+    )
+
+    const errorFileNames = screen.getByTestId("stUploadedFileErrorMessage")
+    expect(errorFileNames.textContent).toContain(
       "text/plain files are not allowed."
     )
-
-    // If a file has an error, the FileUploader still has a valid status
-    expect(instance.status).toBe("ready")
   })
 
-  it("shows an ErrorStatus when maxUploadSizeMb = 0", () => {
+  it("shows an ErrorStatus when maxUploadSizeMb = 0", async () => {
+    const user = userEvent.setup()
     const props = getProps({ maxUploadSizeMb: 0 })
-    const wrapper = shallow(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
-    const fileDropzone = wrapper.find(FileDropzone)
-    fileDropzone
-      .props()
-      .onDrop([], [{ file: createFile(), errors: [FILE_TOO_LARGE] }])
-    const files: UploadFileInfo[] = wrapper.state("files")
+    render(<FileUploader {...props} />)
 
-    expect(files[0].status.type).toBe("error")
-    expect((files[0].status as ErrorStatus).errorMessage).toBe(
+    const fileDropZoneInput = screen.getByTestId(
+      "stDropzoneInput"
+    ) as HTMLInputElement
+
+    await user.upload(fileDropZoneInput, createFile())
+
+    const errorFileNames = screen.getByTestId("stUploadedFileErrorMessage")
+    expect(errorFileNames.textContent).toContain(
       "File must be 0.0B or smaller."
     )
-
-    // If a file has an error, the FileUploader still has a valid status
-    expect(instance.status).toBe("ready")
-  })
-
-  it("does not reset on disconnect", () => {
-    const props = getProps()
-    const wrapper = shallow(<FileUploader {...props} />)
-    const instance = wrapper.instance() as FileUploader
-    // @ts-expect-error
-    const resetSpy = jest.spyOn(instance, "reset")
-    wrapper.setProps({ disabled: true })
-    expect(resetSpy).not.toHaveBeenCalled()
   })
 
   it("resets its value when form is cleared", async () => {
+    const user = userEvent.setup()
+
     // Create a widget in a clearOnSubmit form
     const props = getProps({ formId: "form" })
     jest.spyOn(props.widgetMgr, "setFileUploaderStateValue")
@@ -595,22 +589,26 @@ describe("FileUploader widget enzyme tests", () => {
 
     jest.spyOn(props.widgetMgr, "setIntValue")
 
-    const wrapper = shallow(<FileUploader {...props} />)
+    const { rerender } = render(<FileUploader {...props} />)
+
+    const fileDropZoneInput = screen.getByTestId(
+      "stDropzoneInput"
+    ) as HTMLInputElement
 
     // Upload a single file
-    const fileDropzone = wrapper.find(FileDropzone)
-    fileDropzone.props().onDrop([createFile()], [])
+    await user.upload(fileDropZoneInput, createFile())
 
-    // Wait two ticks so that fileUrls can be fetched and our file uploaded.
-    await process.nextTick
-    await process.nextTick
+    const fileName = screen.getByTestId("stUploadedFile")
+    expect(fileName.textContent).toContain("filename.txt")
 
-    let files = getFiles(wrapper)
-    expect(files).toHaveLength(1)
     expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledWith(
       props.element,
       buildFileUploaderStateProto([
-        (files[0].status as UploadedStatus).fileUrls,
+        {
+          fileId: "filename.txt",
+          uploadUrl: "filename.txt",
+          deleteUrl: "filename.txt",
+        },
       ]),
       {
         fromUi: true,
@@ -619,12 +617,14 @@ describe("FileUploader widget enzyme tests", () => {
 
     // "Submit" the form
     props.widgetMgr.submitForm("form")
-    wrapper.update()
+    rerender(<FileUploader {...props} />)
 
     // Our widget should be reset, and the widgetMgr should be updated
-    files = getFiles(wrapper)
-    expect(files).toHaveLength(0)
-    expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenCalledWith(
+    const fileNames = screen.queryAllByTestId("stUploadedFile")
+    expect(fileNames.length).toBe(0)
+
+    // WidgetStateManager will still have been called once, during component mounting
+    expect(props.widgetMgr.setFileUploaderStateValue).toHaveBeenLastCalledWith(
       props.element,
       buildFileUploaderStateProto([]),
       {
